Update tests for new overview page
