+++ tests/e2e/pages/kubewarden.page.ts
@@ -47,6 +47,10 @@ export class KubewardenPage extends BasePage {
       return this.getPane(pane).locator('span.numbers-stats')
     }
 
+    getPolicyServer(name: string|RegExp) {
+      return this.getPane('Policy Servers').getByRole('listitem').filter({has: this.page.getByRole('link', {name: name, exact: true})})
+    }
+
     @step
     async getCurrentVersion(): Promise<AppVersion> {
       await this.nav.kubewarden()
